% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_MrBean.R
\name{app_MrBean}
\alias{app_MrBean}
\title{User interface for MrBean}
\usage{
app_MrBean()
}
\value{
shiny interface
}
\description{
User interface for MrBean
}
\examples{
# app_MrBean()
}
\author{
Johan Aparicio, \email{j.aparicio@cgiar.org}
}
