% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrBean.R
\name{mrBean}
\alias{mrBean}
\title{User interface for MrBean}
\usage{
mrBean()
}
\value{
shiny interface
}
\description{
User interface for MrBean
}
\examples{
# mrBean()
}
\author{
Johan Aparicio, \email{j.aparicio@cgiar.org}
}
